requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
decorator==5.1.1
Django==4.2.3
django-braces==1.15.0
django-livereload-server==0.4
django-model-utils==4.3.1
django-star-ratings==0.9.2
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
humanize==4.7.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jmespath==1.0.1
lxml==4.9.3
matplotlib-inline==0.1.6
packaging==23.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
Protego==0.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.2.0
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
swapper==1.3.0
tldextract==3.4.4
tornado==6.3.2
traitlets==5.9.0
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.4
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==6.0
